Fix to BaseClient signing process. QoL changes in TryFindProgramAddress nonce return.

+++ src/Solnet.Programs/TokenSwap/TokenSwapProgram.cs
@@ -62,7 +62,7 @@ namespace Solnet.Programs.TokenSwap
         {
             if (!PublicKey.TryFindProgramAddress(new[] { tokenSwapAccount.KeyBytes }, ProgramIdKey, out var auth, out var nonce))
                 throw new InvalidProgramException();
-            return (auth, (byte)nonce);
+            return (auth, nonce);
         }
 
         /// <summary>
